Adding some documentation to the JFXSnackbar class, #932
* I've added documentation to JFXSnackbar as well as SnackbarEvent. Any
feedback for whether it's correct or could use some tweaking is welcome.

* I'm not quite sure what the pseudoClass is used for; could someone more
knowledgeable fill that in?

* JFXSnackbarLayout could use some documentation.

* I refactored the default duration into a constant; hopefully that's
okay. Would it be better suited as a static field?

* Using the javadoc format to delineate sections was causing some weird
javadoc to appear for whatever method followed the section comment;
instead of that format I've used IntelliJ's `sbc` LiveTemplate. It
seems to avoid that problem while maintaining a similar format.