Support tooltips for non-controls
Before this commit, `ValidatorBase` didn't put the error tooltip on non-controls. 

`Tooltip.install(...)` allows one to install a `Tooltip` on any arbitrary node. `Tooltip.install(...)` stores the tooltip in the `Node`'s properties. 

This commit should allow the `ValidatorBase` to add a tooltip to any old `Node`, not just `Control`s—as well as restore the original tooltip, if there was one.