Adds some clarifying documentation for a few things.
My motivation for changing the section comment style is that the old one
makes the IDE think that said block is the javadoc for whatever follows
it.

This new style of section comment is easy to replicate and doesn't
need the right margin fixed if the contents change size. As well as being
easy to repliate for anyone, it also happens to be easy to make in
IntelliJ by using the `sbc` (section block comment) LiveTemplate.